requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
attrs==25.4.0
django==4.2
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
faker==40.5.1
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
psycopg2-binary==2.9.10
pyyaml==6.0.3
redis==7.2.1
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing-extensions==4.15.0
uritemplate==4.2.0
